// src/app/pages/hotel/layout/bed-info/shared/bed-info-enum.ts
export enum  BedInfoEnum {

    guestsOccupancyMaxSize = 10,
    roomPeopleQuantityDefaultValue = 1,
    itemPeopleQuantityDefaultValue = 1,
    bedTypeDefaultValue = 5,
    bedTypeQuatityModelDefaultValue = 1,
    
    roomPeopleQuantity = 'roomPeopleQuantity',
    bedMeasures = 'bedMeasures',
    bedMeasure = 'bedMeasure',
    measureItem = 'measureItem',
    bedQuantity = 'bedQuantity',
    guestsOccupancy = 'guestsOccupancy',
    emptyString = '',
    measureItemDefaultSelection = 1, 


    maxBedTypeInARoom = 12,
    bedMeasuresMaxItems = 3,


    selectUndefinedOptionValue = '',

    getBedMeasuresErrorMessage = 'There was an error trying to get Bed and SofaBed Measures List information',
}